+++ src/InfrastructureModels.jl
@@ -24,6 +24,8 @@ function logger_config!(level; kwargs...)
     Memento.config!(_LOGGER, level, kwargs...)
 end
 
+const nw_id_default = 0
+
 include("core/base.jl")
 include("core/data.jl")
 include("core/constraint.jl")