+++ src/core/base.jl
@@ -32,8 +32,6 @@ InfrastructureModels.@def im_fields begin
     sol::Dict{Symbol,<:Any}
     sol_proc::Dict{Symbol,<:Any}
 
-    cnw::Int
-
     # Extension dictionary
     # Extensions should define a type to hold information particular to
     # their functionality, and store an instance of the type in this
@@ -57,9 +55,6 @@ function InitializeInfrastructureModel(
     sol = _initialize_dict_from_ref(ref)
     sol_proc = _initialize_dict_from_ref(ref)
 
-    # Compute the minimum multinetwork index and store it as `cnw`.
-    cnw = minimum(minimum(collect(keys(ref[:it][it][:nw])) for it in keys(ref[:it])))
-
     imo = InfrastructureModel(
         jump_model,
         data,
@@ -70,7 +65,6 @@ function InitializeInfrastructureModel(
         con,
         sol,
         sol_proc,
-        cnw,
         ext
     )
 
@@ -92,9 +86,6 @@ function InitializeInfrastructureModel(
     sol = _initialize_dict_from_ref(ref)
     sol_proc = _initialize_dict_from_ref(ref)
 
-    # Compute the minimum multinetwork index and store it as `cnw`.
-    cnw = minimum(minimum(keys(ref[:it][it][:nw]) for it in keys(ref[:it])))
-
     imo = InfrastructureModel(
         jump_model,
         data,
@@ -105,7 +96,6 @@ function InitializeInfrastructureModel(
         con,
         sol,
         sol_proc,
-        cnw,
         ext
     )
 
@@ -280,7 +270,7 @@ nw_ids(aim::AbstractInfrastructureModel, it::Symbol) = keys(aim.ref[:it][it][:nw
 nws(aim::AbstractInfrastructureModel, it::Symbol) = aim.ref[:it][it][:nw]
 
 ids(aim::AbstractInfrastructureModel, it::Symbol, nw::Int, key::Symbol) = keys(aim.ref[:it][it][:nw][nw][key])
-ids(aim::AbstractInfrastructureModel, it::Symbol, key::Symbol; nw::Int=aim.cnw) = keys(aim.ref[:it][it][:nw][nw][key])
+ids(aim::AbstractInfrastructureModel, it::Symbol, key::Symbol; nw::Int=nw_id_default) = keys(aim.ref[:it][it][:nw][nw][key])
 
 
 ref(aim::AbstractInfrastructureModel, it::Symbol, nw::Int) = aim.ref[:it][it][:nw][nw]
@@ -289,31 +279,31 @@ ref(aim::AbstractInfrastructureModel, it::Symbol, nw::Int, key::Symbol, idx) = a
 ref(aim::AbstractInfrastructureModel, it::Symbol, nw::Int, key::Symbol, idx, param::String) = aim.ref[:it][it][:nw][nw][key][idx][param]
 
 
-ref(aim::AbstractInfrastructureModel, it::Symbol; nw::Int=aim.cnw) = aim.ref[:it][it][:nw][nw]
-ref(aim::AbstractInfrastructureModel, it::Symbol, key::Symbol; nw::Int=aim.cnw) = aim.ref[:it][it][:nw][nw][key]
-ref(aim::AbstractInfrastructureModel, it::Symbol, key::Symbol, idx; nw::Int=aim.cnw) = aim.ref[:it][it][:nw][nw][key][idx]
-ref(aim::AbstractInfrastructureModel, it::Symbol, key::Symbol, idx, param::String; nw::Int=aim.cnw) = aim.ref[:it][it][:nw][nw][key][idx][param]
+ref(aim::AbstractInfrastructureModel, it::Symbol; nw::Int=nw_id_default) = aim.ref[:it][it][:nw][nw]
+ref(aim::AbstractInfrastructureModel, it::Symbol, key::Symbol; nw::Int=nw_id_default) = aim.ref[:it][it][:nw][nw][key]
+ref(aim::AbstractInfrastructureModel, it::Symbol, key::Symbol, idx; nw::Int=nw_id_default) = aim.ref[:it][it][:nw][nw][key][idx]
+ref(aim::AbstractInfrastructureModel, it::Symbol, key::Symbol, idx, param::String; nw::Int=nw_id_default) = aim.ref[:it][it][:nw][nw][key][idx][param]
 
 var(aim::AbstractInfrastructureModel, it::Symbol, nw::Int) = aim.var[:it][it][:nw][nw]
 var(aim::AbstractInfrastructureModel, it::Symbol, nw::Int, key::Symbol) = aim.var[:it][it][:nw][nw][key]
 var(aim::AbstractInfrastructureModel, it::Symbol, nw::Int, key::Symbol, idx) = aim.var[:it][it][:nw][nw][key][idx]
 
-var(aim::AbstractInfrastructureModel, it::Symbol; nw::Int=aim.cnw) = aim.var[:it][it][:nw][nw]
-var(aim::AbstractInfrastructureModel, it::Symbol, key::Symbol; nw::Int=aim.cnw) = aim.var[:it][it][:nw][nw][key]
-var(aim::AbstractInfrastructureModel, it::Symbol, key::Symbol, idx; nw::Int=aim.cnw) = aim.var[:it][it][:nw][nw][key][idx]
+var(aim::AbstractInfrastructureModel, it::Symbol; nw::Int=nw_id_default) = aim.var[:it][it][:nw][nw]
+var(aim::AbstractInfrastructureModel, it::Symbol, key::Symbol; nw::Int=nw_id_default) = aim.var[:it][it][:nw][nw][key]
+var(aim::AbstractInfrastructureModel, it::Symbol, key::Symbol, idx; nw::Int=nw_id_default) = aim.var[:it][it][:nw][nw][key][idx]
 
 
 con(aim::AbstractInfrastructureModel, it::Symbol, nw::Int) = aim.con[:it][it][:nw][nw]
 con(aim::AbstractInfrastructureModel, it::Symbol, nw::Int, key::Symbol) = aim.con[:it][it][:nw][nw][key]
 con(aim::AbstractInfrastructureModel, it::Symbol, nw::Int, key::Symbol, idx) = aim.con[:it][it][:nw][nw][key][idx]
 
-con(aim::AbstractInfrastructureModel, it::Symbol; nw::Int=aim.cnw) = aim.con[:it][it][:nw][nw]
-con(aim::AbstractInfrastructureModel, it::Symbol, key::Symbol; nw::Int=aim.cnw) = aim.con[:it][it][:nw][nw][key]
-con(aim::AbstractInfrastructureModel, it::Symbol, key::Symbol, idx; nw::Int=aim.cnw) = aim.con[:it][it][:nw][nw][key][idx]
+con(aim::AbstractInfrastructureModel, it::Symbol; nw::Int=nw_id_default) = aim.con[:it][it][:nw][nw]
+con(aim::AbstractInfrastructureModel, it::Symbol, key::Symbol; nw::Int=nw_id_default) = aim.con[:it][it][:nw][nw][key]
+con(aim::AbstractInfrastructureModel, it::Symbol, key::Symbol, idx; nw::Int=nw_id_default) = aim.con[:it][it][:nw][nw][key][idx]
 
 
 sol(aim::AbstractInfrastructureModel, it::Symbol, nw::Int, args...) = _sol(aim.sol[:it][it][:nw][nw], args...)
-sol(aim::AbstractInfrastructureModel, it::Symbol, args...; nw::Int=aim.cnw) = _sol(aim.sol[:it][it][:nw][nw], args...)
+sol(aim::AbstractInfrastructureModel, it::Symbol, args...; nw::Int=nw_id_default) = _sol(aim.sol[:it][it][:nw][nw], args...)
 
 function _sol(sol::Dict, args...)
     for arg in args